requirements.txt
anthropic>=0.49.0
gradio>=4.44.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0
